api/services/qdrant_client.py
# Placeholder for Qdrant client logic
import os
from qdrant_client import QdrantClient
from dotenv import load_dotenv

load_dotenv()

class QdrantService:
    def __init__(self):
        self.qdrant_url = os.getenv("QDRANT_URL")
        self.qdrant_api_key = os.getenv("QDRANT_API_KEY")
        self.client = QdrantClient(
            url=self.qdrant_url,
            api_key=self.qdrant_api_key,
        )
        self.collection_name = "textbook_content"

    def search(self, vector):
        # This is where you would search for similar vectors in Qdrant
        print("Searching in Qdrant...")
        # hits = self.client.search(
        #     collection_name=self.collection_name,
        #     query_vector=vector,
        #     limit=5,
        # )
        # return hits
        return [{"payload": {"text": "mock text from qdrant"}}]

qdrant_service = QdrantService()
