requirements.txt
fastapi
uvicorn[standard]
qdrant-client
psycopg2-binary
python-dotenv
